backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo==4.6.3
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
bcrypt==4.1.1
pyjwt==2.12.1
httpx==0.25.2
razorpay==1.4.1
cloudinary==1.36.0
python-multipart==0.0.6
reportlab==4.2.2
pytest==8.3.3
email-validator==2.2.0
setuptools==80.9.0
prometheus-fastapi-instrumentator==7.1.0
